fix test_create of test_datasources functionnal tests
windows was not able to read the same file many times - it leads to raise
an exception.

use mkstemp to store the file to avoid issues with NamedTemporaryFile
that can't be open a second time on Windows. See:
http://docs.python.org/library/tempfile.html#tempfile.NamedTemporaryFile
"Whether the name can be used to open the file a second time, while the
named temporary file is still open, varies across platforms (it can be so
used on Unix; it cannot on Windows NT or later).